Rename and move game extensions
